refactor: Separate initialization of VideoManager based on test value
The code changes separate the initialization of the VideoManager class based on the value of the 'test' query parameter. If the value is 'True', the video source is set to 'test.jpeg' and displayed on the page. Otherwise, the video stream is started from the device's webcam. This improves flexibility and allows for easier testing.

This commit message follows the established convention of using the "refactor" prefix to indicate code refactoring.